feat: 整理 createApp 结构,对 core/utils 的函数进行整理
