+++ packages/core/src/utils/createApp.ts
@@ -10,21 +10,20 @@ import { removeDirectory, getNpmPackage } from "./fileController";
 import { ProjectTypes, PackageManagers } from "./questions";
 import { projectLink } from "./constants";
 import isGitInstalled from "./checkGitInstallation";
-import createPackageJson from "./createPackageJsonFile";
 import createSuccessInfo from "./createSuccessInfo";
-import createCommitlint from "./createCommitlint";
-import createTemplateFile from "./createTemplateFile";
+import createCommitLint from "./createCommitLint";
+import { createPackageJson, createTemplateFile } from "./createFile";
 
-export default async function createApp(matter: string, options: { force: boolean }) {
-  intro(chalk.green(" create-you-app "));
+// 创建项目文件
+const makeDirectory = async (matter, options) => {
   const rootDirectory = resolveApp(matter);
-
   // 如果存在同名文件,且没有输入 -f,
   if (fs.existsSync(rootDirectory) && !options.force) {
     const shouldContinue = await confirm({
       message:
         "Whether to overwrite a file with the same name that exists in the current directory ?",
     });
+
     // 删除已存在文件并创建新文件
     if (shouldContinue === true) {
       removeDirectory(matter, true);
@@ -32,7 +31,10 @@ export default async function createApp(matter: string, options: { force: boolea
   }
 
   execSync(`mkdir ${rootDirectory}`);
+};
 
+// 获取表单结果
+const getTableInfo = async () => {
   const projectType = (await select({
     message: "Pick a project type.",
     options: ProjectTypes,
@@ -43,23 +45,39 @@ export default async function createApp(matter: string, options: { force: boolea
     options: PackageManagers,
   })) as string;
 
-  const commitlint = (await confirm({
+  const commitLint = (await confirm({
     message: "Pick additional lint features:",
   })) as boolean;
+
+  return { projectType, packageManageType, commitLint }
+}
+
+export default async function createApp(matter: string, options: { force: boolean }) {
+  intro(chalk.green(" create-you-app "));
+  const rootDirectory = resolveApp(matter);
+
+  await makeDirectory(matter, options);
+
+  const { projectType, packageManageType, commitLint } = await getTableInfo()
 
   // 写入 package.json 文件
   fs.writeFileSync(
     join(rootDirectory, "package.json"),
     JSON.stringify(createPackageJson(projectType, matter), null, 2),
   );
+
   // 写入 .gitignore 文件
   fs.writeFileSync(join(rootDirectory, ".gitignore"), createTemplateFile("gitignore"));
 
-  // 下载 npm 包解压,并删除一些无用的代码文件
+  // 下载 npm 包解压,获取目标模板导入文件,并删除一些无用的代码文件
   getNpmPackage(projectLink.get(projectType) as string, projectType, rootDirectory);
 
-  if (commitlint === true) createCommitlint(rootDirectory);
+  // 注入 lint 规则
+  if (commitLint === true) {
+    createCommitLint(rootDirectory);
+  }
 
+  // 安装相关依赖，暂时不用
   const spinner = ora().start();
   spinner.start(chalk.bold.cyan("The dependency package is being installed..."));
   exec(`${packageManageType} install`, { cwd: rootDirectory }, () => {
@@ -69,5 +87,8 @@ export default async function createApp(matter: string, options: { force: boolea
   });
 
   // 是否安装已经安装了 git
-  if (isGitInstalled()) exec("git init", { cwd: rootDirectory });
+  if (isGitInstalled()) {
+    exec("git init", { cwd: rootDirectory });
+  }
 }
+
