--- packages/core/src/utils/createTemplateFile.ts
@@ -1,6 +1,0 @@
-import { readFileSync } from "node:fs";
-import { join } from "node:path";
-
-export default function createTemplateFile(file: string) {
-  return readFileSync(join(__dirname, "../../template/", file)).toString();
-}
